+++ homeassistant/components/smartthings/light.py
@@ -346,9 +346,7 @@ class SmartThingsLamp(SmartThingsEntity, LightEntity):
         # remove 'off' for brightness mapping
         if "off" in levels:
             levels = [level for level in levels if level != "off"]
-        level = percentage_to_ordered_list_item(
-            levels, int(round(brightness * 100 / 255))
-        )
+        level = percentage_to_ordered_list_item(levels, round(brightness * 100 / 255))
         await self.execute_device_command(
             Capability.SAMSUNG_CE_LAMP,
             Command.SET_BRIGHTNESS_LEVEL,
